+++ runtime/runtime/src/mips.rs
@@ -650,6 +650,7 @@ mod tests {
         type Proposal = Call;
         type AddSignerMultiSigTarget = Test;
         type KYCServiceProviders = Test;
+        type Balances = balances::Module<Test>;
     }
 
     impl group::GroupTrait for Test {
