+++ runtime/runtime/src/test/mod.rs
@@ -1,7 +1,11 @@
 pub mod storage;
 pub use storage::TestStorage;
 
+pub mod ext_builder;
+pub use ext_builder::ExtBuilder;
+
 mod asset_test;
 mod identity_test;
 mod multisig;
 mod statistics_test;
+// mod balances_test;
